pyproject.toml
[tool.poetry]
name = "risex-ai-bot"
version = "0.1.0"
description = "AI trading bot for RISE testnet with persona-driven strategies"
authors = ["RISE AI Bot Team"]

[tool.poetry.dependencies]
python = "^3.11"
# Web framework
fastapi = {extras = ["standard"], version = "^0.112.0"}
uvicorn = "^0.30.0"
# HTTP and async
httpx = "^0.25.0"
# Data and configuration
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
# Ethereum and crypto
eth-account = "^0.10.0"
eth-utils = "^4.0.0"
eth-abi = "^5.0.0"
# Utilities
tenacity = "^8.2.3"
loguru = "^0.7.2"
openai = "^2.14.0"
web3 = ">=6.0,<7.0"
aiofiles = "^25.1.0"
tiktoken = "^0.12.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
mypy = "^1.7.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
types-requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "UP", "B", "A", "COM", "C4", "Q", "ARG", "PTH", "ERA", "RUF"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]